# make DateTime.UtcNow 5% faster to minimize the leap second performance impact