Today I've again profiled #25728 and was looking for some ways of lowering the regression.

I started with reading the source code of `FileTimeToSystemTime` and `RtlpTimeToTimeFields` methods. The `RtlpTimeToTimeFields` takes care of leap second support, it knows all leap seconds and knows how to handle the positive and negative ones. Copying this logic from OS to .NET makes obviously no sense, it's too complex and too expensive to maintain.

By inlining some small methods that are called by `DateTime.UtcNow` I was able to make it 5% faster. It ain't much, but that's all I could do.

Before:

![obraz](https://user-images.githubusercontent.com/6011991/62635135-6e841300-b937-11e9-8310-8026e1f9b629.png)

After:

![obraz](https://user-images.githubusercontent.com/6011991/62635236-9f644800-b937-11e9-8058-848fb9bdcc95.png)

```cs
[Benchmark]
public DateTimeOffset GetUtcNow() => DateTimeOffset.UtcNow;

[Benchmark]
public DateTime GetUtcNow() => DateTime.UtcNow;
```


|                Type |    Method |           Toolchain |     Mean | Ratio |
|-------------------- |---------- |-------------------- |---------:|------:|
|       Perf_DateTime | GetUtcNow |  \after\CoreRun.exe | 67.25 ns |  0.95 |
|       Perf_DateTime | GetUtcNow | \before\CoreRun.exe | 70.75 ns |  1.00 |
|                     |           |                     |          |       |
| Perf_DateTimeOffset | GetUtcNow |  \after\CoreRun.exe | 80.81 ns |  0.97 |
| Perf_DateTimeOffset | GetUtcNow | \before\CoreRun.exe | 83.27 ns |  1.00 |

And yes, all the `[MethodImpl(MethodImplOptions.AggressiveInlining)]` were really needed. Moving the throws to `ThrowHelper` was not enough to get those methods inlined.

/cc @danmosemsft