@adamsitnik 2019-08-08
Thanks for all the suggested micro-optimizations, I was able to improve the time by 0.5%