@danmosemsft 2019-08-08
Given the regression from 2.2 was so much larger, my assumption is that it is not worth porting this into 3.0.